Added custom nodeInfo type on response, and query for NodeInfo in order to get binary version. Refactored code to remove circular dependency.
