+++ client/rpc/status.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/cosmos/cosmos-sdk/client/grpc/tmservice"
 	"github.com/tendermint/tendermint/libs/bytes"
 	"github.com/tendermint/tendermint/p2p"
 	coretypes "github.com/tendermint/tendermint/rpc/core/types"
@@ -26,9 +27,21 @@ type validatorInfo struct {
 // ResultStatus is node's info, same as Tendermint, except that we use our own
 // PubKey.
 type resultStatus struct {
-	NodeInfo      p2p.DefaultNodeInfo
+	NodeInfo      nodeInfo
 	SyncInfo      coretypes.SyncInfo
 	ValidatorInfo validatorInfo
+}
+
+type nodeInfo struct {
+	ProtocolVersion p2p.ProtocolVersion
+	DefaultNodeID   p2p.ID
+	ListenAddr      string
+	Network         string
+	Version         string
+	Channels        bytes.HexBytes
+	Moniker         string
+	Other           p2p.DefaultNodeInfoOther
+	BinaryVersion   string
 }
 
 // StatusCommand returns the command to return the status of the network.
@@ -38,6 +51,12 @@ func StatusCommand() *cobra.Command {
 		Short: "Query remote node for status",
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			clientCtx, err := client.GetClientQueryContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			queryClient := tmservice.NewServiceClient(clientCtx)
+			res, err := queryClient.GetNodeInfo(context.Background(), &tmservice.GetNodeInfoRequest{})
 			if err != nil {
 				return err
 			}
@@ -52,8 +71,24 @@ func StatusCommand() *cobra.Command {
 			if err != nil {
 				return err
 			}
+
+			infoWithBinary := nodeInfo{
+				ProtocolVersion: status.NodeInfo.ProtocolVersion,
+				DefaultNodeID:   status.NodeInfo.DefaultNodeID,
+				ListenAddr:      status.NodeInfo.ListenAddr,
+				Network:         status.NodeInfo.Network,
+				Version:         status.NodeInfo.Version,
+				Channels:        status.NodeInfo.Channels,
+				Moniker:         status.NodeInfo.Moniker,
+				Other:           status.NodeInfo.Other,
+				// How do I get the binary version?
+				// This needs to be obtained remotely.
+				// We need a query
+				BinaryVersion: res.ApplicationVersion.GetVersion(),
+			}
+
 			statusWithPk := resultStatus{
-				NodeInfo: status.NodeInfo,
+				NodeInfo: infoWithBinary,
 				SyncInfo: status.SyncInfo,
 				ValidatorInfo: validatorInfo{
 					Address:     status.ValidatorInfo.Address,
